# apiGateway/api/docs/swagger.yaml
definitions:
...
  models.GetUserMoneyResponse:
    properties:
      categoryId:
        type: string
      time:
        type: string
      totalAmount:
        type: string
    type: object
...
  models.Response:
    properties:
      data: {}
      description:
        type: string
      statusCode:
        type: integer
    type: object
  models.Transaction:
    properties:
      accountId:
        type: string
      amount:
        type: number
      categoryId:
        type: string
      createdAt:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      typed:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  models.TransactionsResponse:
    properties:
      count:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/models.Transaction'
        type: array
    type: object
...
  models.UpdateTransaction:
    properties:
      accountId:
        type: string
      amount:
        type: number
      categoryId:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      typed:
        type: string
      userId:
        type: string
    type: object
  user.ProfileResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      fullname:
        type: string
      username:
        type: string
    type: object
  user.ProfileUpdateRequest:
    properties:
      email:
        type: string
      username:
        type: string
    type: object
  user.ProfileUpdateResponse:
    properties:
      email:
        type: string
      fullname:
        type: string
      id:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
  description: Personal Finance Tracker
  title: Personal Finance Tracker API
  version: "1"
paths:
...
  /budget_service/v5/transaction:
    post:
      consumes:
      - application/json
      description: Create a new transaction
      parameters:
      - description: transaction
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/models.Transaction'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new transaction
      tags:
      - transaction
  /budget_service/v5/transaction/{id}:
    delete:
      consumes:
      - application/json
      description: Delete transaction
      parameters:
      - description: transaction_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete transaction
      tags:
      - transaction
    get:
      consumes:
      - application/json
      description: Get transaction by id
      parameters:
      - description: transaction_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get transaction by id
      tags:
      - transaction
    put:
      consumes:
      - application/json
      description: Update transaction
      parameters:
      - description: transaction_id
        in: path
        name: id
        required: true
        type: string
      - description: transaction
        in: body
        name: transaction
        schema:
          $ref: '#/definitions/models.UpdateTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Update transaction
      tags:
      - transaction
  /budget_service/v5/transactions:
    get:
      consumes:
      - application/json
      description: Get all transaction
      parameters:
      - description: page
        in: query
        name: page
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TransactionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all transaction
      tags:
      - transaction
  /budget_service/v5/transactions/income:
    get:
      consumes:
      - application/json
      description: Get all transaction
      parameters:
      - description: user_id
        in: query
        name: user_id
        type: string
      - description: start_time
        in: query
        name: start_time
        type: string
      - description: end_time
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetUserMoneyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all transaction
      tags:
      - transaction
  /budget_service/v5/transactions/spend:
    get:
      consumes:
      - application/json
      description: Get all transaction
      parameters:
      - description: user_id
        in: query
        name: user_id
        type: string
      - description: start_time
        in: query
        name: start_time
        type: string
      - description: end_time
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetUserMoneyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all transaction
      tags:
      - transaction
  /updateprofile:
    put:
      consumes:
      - application/json
      description: Update user Profile
      parameters:
      - description: User update profile request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.ProfileUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.ProfileUpdateResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update user Profile
      tags:
      - auth
  /userprofile/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a user profile by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.ProfileResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: UserProfile a user
      tags:
      - auth
swagger: "2.0"